tracktime add real time spent n double for sessiontime

+++ src/Application/AppToTrack/Abstracts/IAppTimeTracker.cs
@@ -20,14 +20,14 @@ public interface IAppTimeTracker
 	/// </summary>
 	/// <param name="isRunnnig">Indicates whether the appInstance is running at the moment.</param>
 	/// <returns>Current Session Time in Minutes</returns>
-	int GetCurrentSessionTime(bool isRunnnig);
+	double GetCurrentSessionTime(bool isRunnnig);
 
 	/// <summary>
 	/// Get the total AppInstance uptime for selected Date.
 	/// </summary>
 	/// <param name="date">Date value for which to uptime to return.</param>
 	/// <returns>Uptime in minutes.</returns>
-	int GetSessionTimeForDate(DateOnly date);
+	double GetSessionTimeForDate(DateOnly date);
 
 	/// <summary>
 	/// Updates the values in the AppInstance via interactor.
